event_api/pyproject.toml
[project]
name = "event-api"
version = "0.1.0"
description = "Event API"
requires-python = ">=3.12"
dependencies = [
    "aiokafka>=0.12.0",
    "fastapi[standard]>=0.115.6",
    "orjson>=3.10.14",
    "pydantic-settings>=2.7.1",
    "pyjwt[crypto]>=2.10.1",
    "pytest-asyncio>=0.25.2",
    "redis>=5.2.1",
    "user-agents>=2.2.0",
]
